🔧(core) Drop dependency on `tsify-next`
We have made the `ui`-facing wasm API a lot simpler in the last commit,
and the benefits that `tsify-next` provides (mainly exporting enums and
generics) is no longer needed, so we will be only using
`#[wasm-bindgen]` in the future.

This is the last commit in the long train of commits trying to implement
rendering functionalities in `core`. Next we will need to adapt `ui` to
the new API.

Changelog: detail