✨(ui) Split the hex view into components
The hex view was just a single `<div>` with a whole bunch of bytes all
mixed within it. Now it has been better organized into lines of bytes,
which in turn include individual byte `<span>`s.

This makes it feasible to implement more features for the hex view, such
as highlighting byte spans by clicking on packet nodes or vice versa.