🔧 Migrate to `@wasm-tool/rollup-plugin-rust`
This plugin automatically rebuilds `core` so we don't need to manually
ensure artifacts from `core` are not outdated.

This plugin also allows us to pass arbitrary aruguments to
`wasm-bindgen` and other build tools, so we can just pass
`--remove-producers-section` instead of using `wasm-strip`, which saves
us from a lot of hassle and is much more convenient than `wasm-pack`.

Currently the version of `wasm-bindgen` is not pinned, which should be
addressed in a later commit.

Changelog: detail