🧹 Disable `platformAutomerge` for renovate
GitHub's auto-merging is problematic. It does not provide an easy way to
block merging when checks have not been finished, or when checks are
failing.

If you want to prevent auto-merging when `renovate/*` has not passed,
then all other PRs will never be auto-mergable, as they will never have
those checks. And if you want other PRs to be auto-mergable, renovate's
PRs will get auto-merged even when `renovate/*` are failing.

Changelog: skip